Challenge One
A problem to get an answer from user to a multiple choice question and if user inputs the correct answer, print it is correct. if user inputs a wrong answer, print answer is incorrect.
Here, we have to use nextLine() insted of next(). next() can read the input only till the space. It can't read two words separated by a space. nextLine() reads input including space between the words (that is, it reads till the end of line \n ).